+++ jtdx-source/Versions.cmake
@@ -18,6 +18,6 @@
 set (WSJTX_VERSION_MAJOR 1)
 set (WSJTX_VERSION_MINOR 2)
 set (WSJTX_VERSION_32A 0)
-set (WSJTX_VERSION_SUB 0)
+set (WSJTX_VERSION_SUB 1)
 set (WSJTX_RC 0)		 # release candidate number, comment out or zero for development versions
 set (WSJTX_VERSION_IS_RELEASE 1) # set to 1 for final release build